Extract subscription tests to test case
Subscription test case used to test single and all streams subscriptions.
