+++ test/support/shared_storage_test_case.ex
@@ -1,0 +1,17 @@
+defmodule EventStore.SharedStorageTestCase do
+  defmacro define_tests(do: block) do
+    quote do
+      defmacro __using__(options) do
+        block = unquote(Macro.escape(block))
+
+        quote do
+          # use ExUnit.Case
+          use EventStore.StorageCase
+
+          @moduletag unquote(options)
+          unquote(block)
+        end
+      end
+    end
+  end
+end
